fix: Return f64 ratios from grid_ratios to prevent truncation to zero.
Grid placement for floating windows was casting fractional ratios (e.g. 0.5)
to i32 before multiplying by display dimensions, truncating them to 0 and
placing every grid window at the top-left corner.

+++ src/ecs/triggers.rs
@@ -1041,10 +1041,10 @@ fn apply_window_defaults(
     if floating {
         if let Some((rx, ry, rw, rh)) = properties.iter().find_map(WindowParams::grid_ratios) {
             let bounds = active_display.bounds();
-            let x = bounds.width() * rx;
-            let y = bounds.height() * ry;
-            let w = bounds.width() * rw;
-            let h = bounds.height() * rh;
+            let x = (f64::from(bounds.width()) * rx) as i32;
+            let y = (f64::from(bounds.height()) * ry) as i32;
+            let w = (f64::from(bounds.width()) * rw) as i32;
+            let h = (f64::from(bounds.height()) * rh) as i32;
             window.reposition(Origin::new(x, y));
             window.resize(Size::new(w, h), bounds.width());
         }
